+++ zcash_primitives/CHANGELOG.md
@@ -6,6 +6,12 @@ and this library adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `zcash_primitives::transaction::builder`:
+  - `Builder::build` was calling `FeeRule::fee_required` with the number of
+    Sapling outputs that have been added to the builder. It now instead provides
+    the number of outputs that will be in the final Sapling bundle, including
+    any padding.
 
 ## [0.9.0] - 2022-11-12
 ### Added